refactor scaffold
