+++ .github/scripts/manual/send_digest_window-deprecated.py
